### 2024-04-16 #6629 @editorialbot
[PRE REVIEW]: AIgarMIC: a Python package for automated interpretation of agar dilution minimum inhibitory concentration testing
**Submitting author:** @agerada (<a href="http://orcid.org/0000-0002-6743-4271">Alessandro Gerada</a>)
**Repository:** https://github.com/agerada/AIgarMIC
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
