### 2024-04-16 #6629 @agerada
@Kevin-Mattheus-Moerman Dear editor, thank you for considering our submission. I have added contributing guidelines (using a `CONTRIBUTING.md` file in the root), and linked to this in the `README` and project documentation. Also updated UK to United Kingdom. All changes in `main` branch.

From the list of reviewers, I would suggest Dr Griffin Chure (gchure) and Prof Peter MacPherson (petermacp) as potential reviewers.
