### 2024-05-13 #6629 @ppxasjsm
Hi @gchure is there any chance you'd be available to edit this submission to JOSS? Thank you for considering this.
