+++ pyproject.toml
@@ -22,7 +22,7 @@ numpy = ">=1.26"
 opencv-python = ">=4.9"
 pytest = ">=8.1"
 tensorflow-macos = { version = ">=2.11,<=2.15", markers = "platform_system=='Darwin' and platform_machine=='arm64'"}
-tensorflow = { version = ">=2.11,<=2.15", markers = "platform_system!='Darwin' and platform_machine!='arm64'"}
+tensorflow = { version = ">=2.11,<=2.15", markers = "platform_system!='Darwin' and platform_machine!='arm64' or platform_system=='Darwin' and platform_machine!='arm64'"}
 scikit-learn = ">=1.4"
 sphinx = ">=7.2"
 sphinx-argparse = "^0.4.0"
