+++ docs/source/requirements.txt
@@ -1,5 +1,5 @@
 absl-py==2.1.0
--e git+https://github.com/agerada/AIgarMIC.git@60e8c710769c405a4cafba52d0827e197f5248b3#egg=aigarmic
+-e git+https://github.com/agerada/AIgarMIC.git@879dcf0705f7ca802bdf95826e9137da548c7065#egg=aigarmic
 alabaster==0.7.16
 anyascii==0.3.2
 appnope==0.1.4
@@ -25,7 +25,7 @@ executing==2.0.1
 fastjsonschema==2.19.1
 flatbuffers==24.3.25
 fonttools==4.50.0
-gast==0.5.4
+gast==0.4.0
 google-auth==2.29.0
 google-auth-oauthlib==1.2.0
 google-pasta==0.2.0
@@ -48,13 +48,13 @@ jsonschema-specifications==2023.12.1
 jupyter-cache==1.0.0
 jupyter_client==8.6.1
 jupyter_core==5.7.2
-keras==2.15.0
+keras==2.11.0
 kiwisolver==1.4.5
 libclang==18.1.1
 Markdown==3.6
 markdown-it-py==3.0.0
 MarkupSafe==2.1.5
-matplotlib==3.8.3
+matplotlib==3.6.3
 matplotlib-inline==0.1.6
 mccabe==0.7.0
 mdit-py-plugins==0.4.0
@@ -79,7 +79,7 @@ pip-chill==1.0.3
 platformdirs==4.2.0
 pluggy==1.4.0
 prompt-toolkit==3.0.43
-protobuf==4.25.3
+protobuf==3.19.6
 psutil==5.9.8
 ptyprocess==0.7.0
 pure-eval==0.2.2
@@ -116,7 +116,7 @@ sphinxcontrib-serializinghtml==1.1.10
 SQLAlchemy==2.0.29
 stack-data==0.6.3
 tabulate==0.9.0
-tensorboard==2.15.2
+tensorboard==2.11.2
 tensorboard-data-server==0.7.2
 tensorboard-plugin-wit==1.8.1
 tensorflow-estimator==2.15.0
