+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "aigarmic"
-version = "1.0.0"
+version = "1.0.1"
 description = "Machine-learning assisted agar dilution MIC"
 authors = ["Alessandro Gerada"]
 license = "GPL-3.0"
