+++ docs/source/requirements.txt
@@ -27,7 +27,7 @@ flatbuffers==24.3.25
 fonttools==4.50.0
 gast==0.4.0
 google-auth==2.29.0
-google-auth-oauthlib==1.2.0
+google-auth-oauthlib
 google-pasta==0.2.0
 grpcio==1.62.1
 h5py==3.10.0
