### 2024-05-30 84925bd
test minimal req docs
