+++ requirements.txt
@@ -1,137 +1,11 @@
-absl-py==2.1.0
--e git+https://github.com/agerada/AIgarMIC.git@879dcf0705f7ca802bdf95826e9137da548c7065#egg=aigarmic
-alabaster==0.7.16
+aigarmic==1.0.1
 anyascii==0.3.2
-appnope==0.1.4
-astroid==3.1.0
-asttokens==2.4.1
-astunparse==1.6.3
-attrs==23.2.0
-Babel==2.14.0
-cachetools==5.3.3
-certifi==2024.2.2
-charset-normalizer==3.3.2
-click==8.1.7
-comm==0.2.2
-contourpy==1.2.0
-coverage==7.4.4
-cycler==0.12.1
-debugpy==1.8.1
-decorator==5.1.1
-dill==0.3.8
-docutils==0.20.1
-exceptiongroup==1.2.0
-executing==2.0.1
-fastjsonschema==2.19.1
-flatbuffers==24.3.25
-fonttools==4.50.0
-gast==0.4.0
-google-auth==2.29.0
-google-auth-oauthlib==1.2.0
-google-pasta==0.2.0
-grpcio==1.62.1
-h5py==3.10.0
-idna==3.6
-imagesize==1.4.1
-importlib_metadata==7.1.0
-importlib_resources==6.4.0
-imutils==0.5.4
-iniconfig==2.0.0
-ipykernel==6.29.3
-ipython==8.18.1
-isort==5.13.2
-jedi==0.19.1
-Jinja2==3.1.3
-joblib==1.3.2
-jsonschema==4.21.1
-jsonschema-specifications==2023.12.1
-jupyter-cache==1.0.0
-jupyter_client==8.6.1
-jupyter_core==5.7.2
-keras==2.11.0
-kiwisolver==1.4.5
-libclang==18.1.1
-Markdown==3.6
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-matplotlib==3.6.3
-matplotlib-inline==0.1.6
-mccabe==0.7.0
-mdit-py-plugins==0.4.0
-mdurl==0.1.2
-ml-dtypes==0.2.0
 mypy-extensions==1.0.0
-myst-nb==1.0.0
-myst-parser==2.0.0
-nbclient==0.10.0
-nbformat==5.10.3
-nest-asyncio==1.6.0
-numpy==1.26.4
-oauthlib==3.2.2
-opencv-python==4.9.0.80
+myst-nb==1.1.0
 opencv-python-headless==4.9.0.80
-opt-einsum==3.3.0
-packaging==24.0
-parso==0.8.3
-pexpect==4.9.0
-pillow==10.2.0
 pip-chill==1.0.3
-platformdirs==4.2.0
-pluggy==1.4.0
-prompt-toolkit==3.0.43
-protobuf==3.19.6
-psutil==5.9.8
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pyasn1==0.5.1
-pyasn1-modules==0.3.0
-Pygments==2.17.2
 pylint==3.1.0
-pyparsing==3.1.2
-pytest==8.1.1
 pytest-cov==5.0.0
-python-dateutil==2.9.0.post0
-PyYAML==6.0.1
-pyzmq==25.1.2
-referencing==0.34.0
-requests==2.31.0
-requests-oauthlib==2.0.0
-rpds-py==0.18.0
-rsa==4.9
-scikit-learn==1.4.1.post1
-scipy==1.12.0
-six==1.16.0
-snowballstemmer==2.2.0
-Sphinx==7.2.6
-sphinx-argparse==0.4.0
-sphinx-autoapi==3.0.0
+sphinx-autoapi==3.1.1
[…]
