### 2024-06-02 #6826 @editorialbot
[REVIEW]: AIgarMIC: a Python package for automated interpretation of agar dilution minimum inhibitory concentration testing
**Submitting author:** @agerada (<a href="http://orcid.org/0000-0002-6743-4271">Alessandro Gerada</a>)
**Repository:** https://github.com/agerada/AIgarMIC
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
