### 2024-06-04 #6826 @gchure
Hi @ppxasjsm, I have a quick editorial question. It seems like the software is [already published](https://journals.asm.org/doi/10.1128/spectrum.04209-23) in another peer reviewed journal ASM Spectrum. From my read of that paper, the software is described and validated on a novel dataset generated solely to train and test the software (i.e., the software isn't used in the paper to facilitate discovery). Given the [JOSS guidelines for copublication](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software), it's not clear to me that this software by itself merits a separate JOSS paper, even though the ASM paper doesn't cover the API or structure of the software itself.

Could you please let me know your opinion on this? I'm happy to continue to review the software (which seems quite good), but I noticed that almost all of my comments on the software paper itself are covered in the ASM publication.
