### 2024-06-04 #6826 @agerada
Thank you @gchure for your comments and for reviewing our software. Although I appreciate the question is for @ppxasjsm, I will leave some comments to help with the decision.

Our JOSS submission for `AIgarMIC` was alongside a technical paper (in ASM Microbiology Spectrum) which we reference in the JOSS draft paper. The primary focus of the ASM Spectrum paper was to report and validate the use of machine learning approaches to interpret minimimum inhibitory concentration (MIC) experiments, delivered through our `AIgarMIC` software. Within the paper, we describe the laboratory experiments that generated the training images, the specifications of the artificial neural networks that were used to label images, and the validation of the predictions by comparing against a gold standard method (manual interpretation). The results are reported in the domain-standard format for MIC validation studies (e.g., essential and categorical agreement). Within the methods, we included a paragraph outlining the high-level features of `AIgarMIC`, to demonstrate a practical implementation of our approach.

Since the focus of the ASM Spectrum paper was the microbiological validation of our methods, we submitted our software to JOSS on the basis that our implementation contains software-specific scientific efforts (which are not explained or discussed in the ASM Spectrum paper). These include, for example:

* the algorithms that split images, calculate the MIC, and perform quality assurance of the results.
* the abstraction of agar dilution plates and experiments into `Plate` and `PlateSet` classes (with respective methods), facilitating user extension.
* the use of a black over- or underlying grid for reliable software splitting of images.
* tutorials that explain how to implement `AIgarMIC` in a laboratory, with different use-cases (ranging from use as an MIC calculator to extension with custom models).
* command-line scripts that allow users without Python scripting experience to generate their own labelled images, train and test models, and perform MIC predictions.
* the testing suite focused on the software implementation and algorithms (which complements the technical validation reported in ASM Spectrum).

We look forward to further comments and again thank @gchure and @ppxasjsm.
