### 2024-06-04 #6826 @gchure
Thanks @agerada, that is really helpful context and I appreciate the effort here. Just to overcommunicate, I'm happy to give it a full JOSS review and just wanted to check with @ppxasjsm as I thought it may be in conflict with the JOSS guidelines.

As to the algorithms for splitting, are there tutorials/derivations of the splitting algorithms? I couldn't find details in the documentation beyond the API reference and source code. It may be worthwhile to write a tutorial outlining in some detail the algorithms and math behind the MIC calculation, for example.
