### 2024-06-04 agerada/AIgarMIC#10 @agerada
Docs - add manual splitting tutorial to advanced topic
Tutorial on how to use splitting algorithms is missing.

In [Advanced usage](https://aigarmic.readthedocs.io/en/latest/advanced.html#advanced-usage) section, add scripting subheader and include tutorial on how to link images to `Plate` class, and explicitly trigger the splitting algorithm.
