### 2024-06-04 agerada/AIgarMIC#11 @agerada
Docs - add technical description of algorithms
Add new page to documentation to provide technical description of the following algorithms used within `AIgarMIC`:

- [x] grid splitting
- [x]  MIC calculation
- [x]  QC control algorithm
