### 2024-06-04 #6826 @agerada
Thank you @gchure – fair points on the documentation and algorithms. I have reviewed the documentation and the tutorials don't include use of the splitting alogirthm, which would be helpful. Also agree an overview of how the MIC and QC algorithms are implemented would be helpful. I will likely do this in pseudocode as the implementations are in software rather than in formal math.

Pending further input from @ppxasjsm I have opened two issues (agerada/AIgarMIC#10 and agerada/AIgarMIC#11 respectively) for the two points raised.
