### 2024-06-07 agerada/AIgarMIC#12 @gchure
JOSS Review - CI/CD for non-asset tests and code test coverage
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/6826) and will be opening issues as I go along through the checklist. This pertains to the test suite for the software.

I was able to install the software following your instructions and was able to run both the asset-based tests and the more minimal test suite. Is there a reason you do not have this running as a [github action](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python#testing-with-pytest-and-pytest-cov) upon push to main? I understand downloading the assets may not be feasible going this route, but any amount of routine testing would be wise for promoting contributions.

Additionally, I think it would be useful to automatically generate some test coverage reports. The testing suite looks very good (although some documentation for the test functions would be nice!), but it's hard for me to see what fraction of the core software is covered by these tests. I have experience using [codecov](https://about.codecov.io/) which I have enjoyed, though there are other options.

Neither of these points are blocking, but are "nice-to-haves".
