### 2024-06-08 agerada/AIgarMIC#12 @agerada
Thank you @gchure for the helpful comments and feedback. I agree with all the points and will try to implement as many of them as I can. Much of the testing approach was driven by a desire to use actual assets (images, models) in the tests. Initially we did not have an assets flag for pytest. Since we now do, we should be fine to have a github action, albeit it will only be for the minimal testing suite.

I'm not familiar with codecov, but it looks very useful. I will look into whether it is compatible with assets, as that would be quite important for this project's code coverage. During development, I used [pytest-cov](https://pypi.org/project/pytest-cov/), which is a project dependency. I had an issue with subprocesses not being measured for coverage, leading to the command line scripts tests not being included in the the report, but I think I have found a solution for this. I will explore this and add as a patch if it works.

- [ ] add github action for "assets not required" tests
- [ ] `pytest-cov` notes in [developer](https://aigarmic.readthedocs.io/en/latest/developer.html#running-tests) section of docs
- [ ] expand test coverage to command line scripts, see: https://stackoverflow.com/a/77107013
- [ ] review codecov compatibility with assets
- [ ] add more documentation/comments to tests
