+++ .github/workflows/pytest.yml
@@ -1,0 +1,32 @@
+name: Run tests
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10.x"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install pytest-cov
+        pip install -e ./ 
+   - name: Test with pytest
+      run: |
+        curl -O https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
+        unzip aigarmic_assets.zip
+        pytest
