+++ .github/workflows/pytest.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10.x"]
+        python-version: ["3.9.x"]
 
     steps:
     - uses: actions/checkout@v3
@@ -25,7 +25,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install pytest-cov
         pip install -e ./ 
-   - name: Test with pytest
+    - name: Test with pytest
       run: |
         curl -O https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
         unzip aigarmic_assets.zip
