+++ .github/workflows/pytest.yml
@@ -27,6 +27,6 @@ jobs:
         pip install -e ./ 
     - name: Test with pytest
       run: |
-        curl -O https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
+        curl -kO https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
         unzip aigarmic_assets.zip
         pytest
