### 2024-06-10 5526aa1
Reworking action using 7z for unzipping archive
