+++ .github/workflows/pytest.yml
@@ -16,6 +16,7 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
+    - uses: milliewalky/setup-7-zip@v1
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
@@ -26,7 +27,6 @@ jobs:
         pip install pytest-cov
         pip install -e ./ 
     - name: Test with pytest
-      uses: milliewalky/setup-7-zip@v1
       run: |
         apt-get install 
         curl -kO https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
