### 2024-06-10 7b8a9c1
Removing install of 7zip as I think it's already present in the runner
