### 2024-06-10 agerada/AIgarMIC#12 @gchure
Hi @agerada, I just submitted a PR to run CI/CD with all of the assets (see #13).
