### 2024-06-10 agerada/AIgarMIC#10 @agerada
Updated [documentation](https://aigarmic.readthedocs.io/en/latest/advanced.html#using-custom-image-splitters) to show how to explicitly call the splitting algorithm, and how to provide a different splitter.

See bb3b2b071be96de222c5851dc73bfe9517c57535

Tagging @gchure for review
