### 2024-06-10 2d92f94
qc added to technical
