+++ docs/source/technical.rst
@@ -10,7 +10,9 @@ The first key step that must be performed reliably is the splitting of an agar p
 
 Secondly, in real-world use, we often encounter stray colonies that would disturb algorithms that anchor to colony growth. For these reasons, we found the use of a black grid on the agar plate as the most practical and reliable method to direct an image splitting algorithm. We applied the grid (using transparent paper) during the photography step, allowing us to adjust the grid to compensate for stray colonies. The grid also worked well in plates with high antimicrobial concentrations, where growth can be sparse.
 
-The algorithm first needs to map out the grid using `simple thresholding <https://docs.opencv.org/4.x/d7/d4d/tutorial_py_thresholding.html>`_. If the grid is added in software using a perfect black, this is easy, since a threshold value of ``0`` should work. If the grid is applied during photography, it may be an off black/grey color. To solve this issue, the algorithm searches for the lowest threshold value that returns the expected number of small images (split using `contours <https://docs.opencv.org/4.x/d4/d73/tutorial_py_contours_begin.html>`_::
+The algorithm first needs to map out the grid using `simple thresholding <https://docs.opencv.org/4.x/d7/d4d/tutorial_py_thresholding.html>`_. If the grid is added in software using a perfect black, this is easy, since a threshold value of ``0`` should work. If the grid is applied during photography, it may be an off black/grey color. To solve this issue, the algorithm searches for the lowest threshold value that returns the expected number of small images (split using `contours <https://docs.opencv.org/4.x/d4/d73/tutorial_py_contours_begin.html>`_: [#f1]_
+
+.. code-block:: Delphi
 
     procedure find_threshold(image, max_threshold, n_rows, n_cols);
     for i := 1 to max_threshold do
@@ -21,7 +23,9 @@ The algorithm first needs to map out the grid using `simple thresholding <https:
         end
     end
 
-Now that we have a threshold value that generates the expected number of contours, we can use the contours to split the image into small images. To make sure each image is in the correct position within the matrix that we generate, we iterate through the contours, sorted by their x-y co-ordinates::
+Now that we have a threshold value that generates the expected number of contours, we can use the contours to split the image into small images. To make sure each image is in the correct position within the matrix that we generate, we iterate through the contours, sorted by their x-y co-ordinates: [#f2]_
+
+.. code-block:: Delphi
 
     procedure split_by_grid(image, contours, n_rows, n_cols);
     image_matrix := array[1..n_rows, 1..n_cols] of image;
@@ -35,19 +39,21 @@ Now that we have a threshold value that generates the expected number of contour
 
 Finally, we end up with an matrix of images, each representing a single inoculation site.
 
-MIC determination algorithm
----------------------------
+MIC determination
+-----------------
 
 Assuming that we have an ``image_matrix`` for each agar plate in our experiment, we can combine these into a 3-dimensional matrix (in ``AIgarMIC`` this would correspond to a :class:`aigarmic.PlateSet` object). Assuming further that each ``image_matrix`` has been converted to a matrix of colony growth codes, the MIC determination algorithm can be applied.
 
-We will assume that colony growth codes are one of ``(0, 1, 2)``, and that the threshold for uninhibited growth is ``2`` (i.e., values under ``2`` will be counted as inhibited growth). To calculate the MIC at position ``x, y``, we will search for the concentration at which growth starts to be inhibited. To do this, we will iterate in reverse until we find a concentration where growth is inhibited::
[…]
