### 2024-06-10 agerada/AIgarMIC#11 @agerada
Added new documentation [page](https://aigarmic.readthedocs.io/en/latest/technical.html#technical-background) to give explanation of original alogirthms used in `AIgarMIC`.

See #15

Tagging @gchure for review - thank you.

I will also run through the pseudo-code again to make sure correct.
