### 2024-06-10 agerada/AIgarMIC#16 @gchure
JOSS Review - Paper Comments
Hi! I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/6826) and am opening issues as I make my way through the checklist. This issue pertains to the software paper.

Overall, I think the paper is good and meets the requirements for JOSS. None of my comments below are blocking, and you should take/leave them at your (and @ppxasjsm) discretion.

## Clarity

* I’m a practicing microbiologist, and even though I live and breathe quantitative plating assays, I had a hard time understanding precisely what AIgarMIC is used for. I think having a diagram showing the basic workflow of a MIC assay would help the reader understand what type of data AIgarMIC is expecting. Repurposing [figure 1 from your ASM Spectrum](https://journals.asm.org/doi/epdf/10.1128/spectrum.04209-23) paper I think would be very helpful (which should be fine, considering it’s CC-BY 4.0 licensed).

* I think having an additional figure showing an API diagram for AIgarMIC would be very helpful. This is necessary beyond developers, so I think it’s worthwhile including in the paper, even though you have pretty detailed documentation.

* As AIgarMIC is at heart a ML model, you should provide some information both in the paper and the documentation about the structure of the network and some of the design decisions you’ve made. You did this in your ASM spectrum paper, but I find its absence from the JOSS paper to be a bit confusing. I strongly think you should include a section outlining the structure of the network and information about how it has been trained, pointing to your ASM paper for more detail on the validation of its efficacy. Additionally, this will allow you to reference libraries such as Keras that you use extensively in the software.

## Statement of Need
The statement of need is pretty sparse and could be more clear. My view is that most MIC is done qualitatively by visual inspection of plates or very rudimentary measurements. AIgarMIC provides a means by which to make this more robust and transferrable between labs and scientists. There should be a brief outline of how this emerges.

## Line-by-line Comments

Here are some minor comments that I think would help the language of the paper. Again, take or leave these as you see fit.

- Line 11: MIC is not a test, it’s a quantity that an assay is performed to infer.
- Line 12: Clinic testing is one of the outcomes, not the only one. It’s broadly used in microbiology.
- Line 17: “[…] bacterial growth on each position on solid […]”. It’s not clear from this sentence what “position” means, unless you are deeply familiar with how these assays are technically run. I would be more pedagogical and verbose with what you mean (or better yet, use a figure).
- Line 18: “[…] 3-dimensional growth matrix.” What are the dimensions? Time, concentration, strain? I think you should be more clear with what types of data AIgarMIC is designed to process.
- Line 21 - 22: “[…] Users can customise AIgarMIC to their workflow with bespoke models, or use the pre-trained models provide.” This is not clear or supported in the paper, though the documentation is detailed. Giving a small example or pointing to specific pages in the documentation would be really useful.
- Line 23: “[…] minimizes measurement error.” This is very broad and does not go into what types of “measurement error” are considered. Minimizing implies an optimization procedure, which I do not think is warranted. I think this should be replaced with “reduces subjective manual assignment error” or something along those lines.
