### 2024-06-12 agerada/AIgarMIC#16 @agerada
Thank you @gchure for the helpful comments. We will review the paper in light of these comments over the next few days, and produce a revised version and response.
