+++ src/aigarmic/plate.py
@@ -22,7 +22,7 @@ from warnings import warn
 class Plate:
     def __init__(self, drug: str,
                  concentration: float,
-                 image_path: Optional[str] = None,
+                 image: Optional[Union[str, cv2.typing.MatLike]] = None,
                  n_row: Optional[int] = None,
                  n_col: Optional[int] = None,
                  growth_code_matrix: Optional[list[list[int]]] = None,
@@ -34,7 +34,7 @@ class Plate:
 
         :param drug: Antibiotic name
         :param concentration: Antibiotic concentration
-        :param image_path: Path to the image
+        :param image: CV2 image array or the path to the image
         :param n_row: Number of rows in the plate
         :param n_col: Number of columns in the plate
         :param visualise_contours: Visualise the contours of the plate (useful for validation of grid splitting)
@@ -43,18 +43,12 @@ class Plate:
         """
         self.drug = drug
         self.concentration = concentration
-        self.image_path = image_path
-
-        if image_path is not None:
-            if n_row is None or n_col is None:
-                raise ValueError("Plate dimensions must be provided if image path is provided")
+        self.image = image
         self.n_row = n_row
         self.n_col = n_col
-
         self.accuracy_matrix = None
         self.model = model
         self.key = None
-
         if key is not None:
             if self.model is not None and self.model.get_key() != key:
                 warn(f"Key provided to Plate does not match linked model key: {key} vs {self.model.get_key()}")
@@ -71,17 +65,20 @@ class Plate:
             self.add_growth_code_matrix(growth_code_matrix)
         else:
             self.growth_code_matrix = None
-
         self.growth_matrix = None
         self.score_matrix = None
         self.predictions_matrix = None
         self.image_matrix = None
-        self.image = None
         self.model_image_x = None
         self.model_image_y = None
-        
-        if self.image_path: 
-            self.image = cv2.imread(self.image_path)  # pylint: disable=no-member
+
+        if self.image is not None:
+            if n_row is None or n_col is None:
+                raise ValueError("Plate dimensions must be provided if image path is provided")
+
+            if isinstance(self.image, str):
+                self.image = cv2.imread(self.image)  # pylint: disable=no-member
+
             self.image_matrix = split_by_grid(self.image, self.n_row, visualise_contours=visualise_contours,
                                               plate_name=self.drug + '_' + str(self.concentration))
 
@@ -229,8 +226,8 @@ class Plate:
                 Unable to find an image_matrix associated with this plate. 
                 Please provide an image path on construction or use import_image and split_images()
                 """)
-        model = self.model if not model else model 
-        if not model: 
+        model = self.model if not model else model
+        if not model:
             raise LookupError(
                 """
                 Unable to find an image model for predictions associated with this plate. 
@@ -247,8 +244,8 @@ class Plate:
         temp_growth_rows = []
         temp_growth_code_rows = []
         temp_accuracy_row = []
-        for row in self.image_matrix: 
-            for image in row: 
+        for row in self.image_matrix:
+            for image in row:
                 prediction_data = model.predict(image)
                 if 'growth_code' not in prediction_data:
                     raise ValueError("Model predictions dict must contain 'growth_code'")
@@ -280,11 +277,11 @@ class Plate:
         """
         Print growth matrix in human-readable format
         """
-        if not self.growth_matrix: 
+        if not self.growth_matrix:
             print(f"Plate {self.drug} - {self.concentration} not annotated")
-        else: 
[…]
