### 2024-06-13 eb88965
get_image_paths update
