### 2024-06-14 9c6db81
main.py extends support beyond 96-position plates
