### 2024-06-18 20e4db5
reorganisation of inheritance API
