+++ paper.md
@@ -38,6 +38,8 @@ bibliography: paper.bib
 
 Minimum inhibitory concentration (MIC) is a laboratory test used to estimate  the susceptibility of a microorganism to an antibiotic. The result is used to determine whether it is possible to use that same drug to treat a patient's infection. Agar dilution is a reference method for MIC testing. However, the interpretation of agar dilution plates is time-consuming and prone to intra- and inter-operational error when read by laboratory personnel. `AIgarMIC` is a Python package for automated interpretation of agar dilution methodology. `AIgarMIC` processes laboratory images to identify bacterial growth on each position on solid agar containing different dilutions of an antimicrobial agent to generate a 3-dimensional growth matrix. The growth matrix is then used to identify the antimicrobial concentration at which microbial growth is inhibited -- defining the minimum inhibitory concentration. `AIgarMIC` can be imported for use in Python scripts, or can be run through a command-line interface. Users can customise `AIgarMIC` to their workflow with bespoke models, or use the pre-trained models provided. `AIgarMIC` automates the collection of multiple data and minimizes measurement error.
 
+![``AIgarMIC`` user API](paper_images/api.png)
+
 # Statement of need
 
 Antimicrobial susceptibility testing (AST) is required to ensure timely and appropriate antimicrobial therapy worldwide. AST is also used to quantify the incidence and prevalence of antimicrobial resistance in hospitals, regions and countries. Agar dilution is a standard AST method -- it has the advantage of being relatively inexpensive, and enables high throughput. However, the implementation of agar dilution is often limited by the time required to interpret plates, a process that is also subject to significant intra- and inter-observer variability.
