+++ paper.bib
@@ -36,11 +36,18 @@
   title = {Image and Model Assets for {{AIgarMIC}} (Agar Dilution Minimum Inhibitory Concentration Software)},
   author = {Gerada, Alessandro and Harper, Nicholas and Howard, Alex and Reza, Nada and Hope, William and {Liverpool Clinical Laboratories}},
   date = {2024},
-  publisher = {[object Object]},
+  publisher = {DataCat},
   doi = {10.17638/DATACAT.LIVERPOOL.AC.UK/2631},
   url = {https://datacat.liverpool.ac.uk/id/eprint/2631},
   urldate = {2024-04-03},
   abstract = {This data repository contains assets (primarily images and trained neural network models) to be used with AIgarMIC:  https://github.com/agerada/AIgarMIC.}
+}
+
+@software{itseez2015opencv,
+  title = {Open Source Computer Vision Library},
+  author = {{Itseez}},
+  date = {2015},
+  url = {https://github.com/itseez/opencv}
 }
 
 @article{lamprechtCellProfilerFreeVersatile2007,
