+++ paper.md
@@ -36,17 +36,60 @@ bibliography: paper.bib
 
 # Summary
 
-Minimum inhibitory concentration (MIC) is a laboratory test used to estimate  the susceptibility of a microorganism to an antibiotic. The result is used to determine whether it is possible to use that same drug to treat a patient's infection. Agar dilution is a reference method for MIC testing. However, the interpretation of agar dilution plates is time-consuming and prone to intra- and inter-operational error when read by laboratory personnel. `AIgarMIC` is a Python package for automated interpretation of agar dilution methodology. `AIgarMIC` processes laboratory images to identify bacterial growth on each position on solid agar containing different dilutions of an antimicrobial agent to generate a 3-dimensional growth matrix. The growth matrix is then used to identify the antimicrobial concentration at which microbial growth is inhibited -- defining the minimum inhibitory concentration. `AIgarMIC` can be imported for use in Python scripts, or can be run through a command-line interface. Users can customise `AIgarMIC` to their workflow with bespoke models, or use the pre-trained models provided. `AIgarMIC` automates the collection of multiple data and minimizes measurement error.
+Minimum inhibitory concentration (MIC) is a laboratory test used to estimate  the susceptibility of a microorganism to an antibiotic. The result is used to determine whether it is possible to use that same drug to treat a patient's infection. Agar dilution is a reference method for MIC testing. However, the interpretation of agar dilution plates is time-consuming and prone to intra- and inter-operational error when read by laboratory personnel. `AIgarMIC` is a Python package for automated interpretation of agar dilution images. 
 
-![``AIgarMIC`` user API](paper_images/api.png)
+![High-level overview of the integration of `AIgarMIC` within the laboratory pathway of minimum inhibitory concentration measurement using agar dilution. `AIgarMIC` performs the interpretative steps of the pathway (from step 5), taking a set of agar plates with a colony-locating grid as an input, and reporting an MIC for each isolate. In this example, 4x4 strains are inoculated onto agar plates, giving a total of 36 strains.\label{fig:overview}](paper_images/overview.pdf)
+
+From an input of agar plate images generated through agar dilution (usually consisting of a positive control plate and multiple plates with serial dilutions of antimicrobial concentration), `AIgarMIC` returns an MIC for each microorganism strain in the experiment. \autoref{fig:overview} provides a high-level overview of how `AIgarMIC` achieves this. Firstly, each agar plate image is split into smaller images for each bacterial strain. Next, using a pre-trained image classification model, the small colony images are converted to a code representing growth level (e.g., good growth, inhibited growth) and stored in a matrix for each plate. Finally, `AIgarMIC` uses the growth matrices from all plates to identify the antimicrobial concentration at which microbial growth is inhibited -- the minimum inhibitory concentration. `AIgarMIC` can be imported for use in Python scripts, or can be run through a command-line interface. Users can customise `AIgarMIC` to their workflow with bespoke models, or use the pre-trained models provided. `AIgarMIC` automates the collection of multiple data and minimizes measurement error.
+
+# Software design
+
[…]
