### 2024-06-19 36dd3c4
added keras ref, minor changes to figure 1
