+++ paper.bib
@@ -1,3 +1,10 @@
+@software{chollet2015keras,
+  title = {Keras},
+  author = {Chollet, François and others},
+  date = {2015},
+  url = {https://keras.io}
+}
+
 @article{eucastDeterminationMinimumInhibitory2000,
   title = {Determination of Minimum Inhibitory Concentrations ({{MICs}}) of Antibacterial Agents by Agar Dilution},
   author = {{EUCAST}},
