### 2024-06-20 dc4e944
remove old pytest statement
