+++ .github/workflows/pytest.yml
@@ -30,11 +30,10 @@ jobs:
         python -m pip install --upgrade pip
         pip install pytest-cov
         pip install -e ./ 
-    - name: Test with pytest
+    - name: Download assets
       run: |
         curl -kO https://datacat.liverpool.ac.uk/2631/2/aigarmic_assets.zip
         7z x aigarmic_assets.zip -aoa
-        pytest
     - name: Build coverage file
       run: |
         pytest --junitxml=pytest.xml --cov-report=term-missing:skip-covered --cov=aigarmic tests/ | tee pytest-coverage.txt
