### 2024-06-20 agerada/AIgarMIC#12 @agerada
Hi @gchure. Just following up with an update on my progress with this issue.

- Firstly, I have added some additional comments for tests to help follow the testing suite a9add4abc8fa1945aff14246adb9eea0620e4174
- Regarding code coverage, I made a small change to your code to generate a coverage report comment on pull requests 6a921ca1a509271b0b74affd1e128d27154237ee. This seems to work quite well and fits well within the current workflow.
- I have also added a [section](https://aigarmic.readthedocs.io/en/latest/developer.html#test-coverage) to the developer section of the documentation explaining test code coverage.

All of these changes are in PR #23.

Thank you again for your helpful feedback and review.
