### 2024-06-20 agerada/AIgarMIC#12 @gchure
Hi, #23  looks great. I think it would be useful to pipe the coverage report to a service like [codecov](https://codecov.io) so the coverage can be easily viewed (like [this](https://app.codecov.io/gh/cremerlab/hplc-py) which I configured for my recent joss paper).

Feel free to veto that as desired, but I think that generating the coverage in a workflow is a little useless unless the report is accessible.
