### 2024-06-20 agerada/AIgarMIC#12 @agerada
Thanks for forwarding your example – I do see the beneift now. I went ahead and set it up (#24), accessible [here](https://app.codecov.io/gh/agerada/AIgarMIC).
