### 2024-06-21 agerada/AIgarMIC#12 @gchure
This is great. Thanks for handling that so swiftly. I'm satisfied with this and will close the issue and continue with the remaining bits of my review soon.
