### 2024-06-21 agerada/AIgarMIC#16 @gchure
I read through the paper and my concerns are assuaged. Thanks for taking care of them and giving them thorough consideration!

I cannot build the pdf using [Inara](https://marketplace.visualstudio.com/items?itemName=dfm.vscode-inara) nor through the [JOSS editorial bot](https://github.com/openjournals/joss-reviews/issues/6826#issuecomment-2181658164), though the error messages aren't too illuminating. Were you able to build the PDF locally?
