### 2024-06-21 agerada/AIgarMIC#16 @agerada
The pdf issue was quite strange as I haven't had any problems building the paper using the JOSS github [action](https://github.com/agerada/AIgarMIC/actions/workflows/draft-pdf.yml). Looking through the error, I wonder whether it was a temporary issue with Inara, as it failed prior to attempting to compile the pdf. I triggerred editorialbot just now and it seems to have [compiled](https://github.com/openjournals/joss-reviews/issues/6826#issuecomment-2183504730) without issue.
