### 2024-06-24 agerada/AIgarMIC#11 @gchure
Thanks for adding this. I want to make sure I understand something correctly with the MIC -- in AIgarMIC, the MIC must be one of the concentrations used in the original screen, correct? Looking through things, I can't find a specific model that you are fitting -- rather, you're iterating through concentrations until you find one where the growth is inhibited.
