### 2024-06-24 #6826 @gchure
Hi @agerada and @ppxasjsm, I can just about close out my review of this work. There's two small outstanding issues(see [#11](https://github.com/agerada/AIgarMIC/issues/11) and [#16](https://github.com/agerada/AIgarMIC/issues/16)), but I don't think they will require further review. Thanks @agerada for being collaborative and prompt with your responses to things (and sorry my review took longer than I had anticipated!).
