+++ docs/source/advanced.rst
@@ -26,10 +26,10 @@ Now we will use :func:`aigarmic.get_paths_from_directory` to get the paths of im
 >>> paths = get_paths_from_directory("../images/antimicrobials/")
 >>> plates = [Plate(antibiotic, get_concentration_from_path(path), model=model, image=path, n_row=8, n_col=12) for path in paths[antibiotic]]
 
-Note also that we provided an (optional) key to ``MeanModel`` on construction, to help with the interpretation of the growth code. The next step involves annotating the images, i.e., converting from images to growth matrices. This step will utilise the linked ``MeanModel`` for each plate:
+Note also that we provided an (optional) key to ``MeanModel`` on construction, to help with the interpretation of the growth code. Since we provided both an image and a valid model to the :class:`aigarmic.Plate` constructor, the agar plate image has been split into smaller colony images, and each colony image converted to a growth code (using the linked model). We can confirm this by inspecting the `growth_code_matrix` (here we will look at the first row of the first plate):
 
->>> for i in plates:
-...     _ = i.annotate_images()
+>>> plates[0].growth_code_matrix[0]
+[0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0]
 
 Finally, we can convert the list of plates to a :class:`aigarmic.PlateSet` object, and calculate the MICs:
 
@@ -57,12 +57,12 @@ A key step in calculating MICs using ``AIgarMIC`` is the splitting of agar plate
 ...     image = cv2.imread(path)
 ...     plate.image_matrix = split_by_grid(image, n_rows=8)
 >>> for i in plates:
-...     _ = i.annotate_images()
+...     _ = i.annotate_images()  # explicit call needed
 >>> plate_set = PlateSet(plates_list=plates)
 >>> _ = plate_set.calculate_mic(no_growth_key_items=tuple([0]))
 >>> plate_set.convert_mic_matrix(mic_format='string').tolist()[0]
 ['32.0', '32.0', '32.0', '>64.0', '>64.0', '>64.0', '64.0', '32.0', '16.0', '16.0', '16.0', '16.0']
 
-The code here is similar to previous examples, except that we are not passing an image path on initialisation of :class:`aigarmic.Plate`. Instead, we are reading the image (using ``opencv``) and splitting (using :func:`aigarmic.process_plate_image.split_by_grid`) into small images. The key point here is that we then overwrite the attribute ``image_matrix`` of the :class:`aigarmic.Plate` object with the split images. From this point onwards, the code is the same as before.
+The code here is similar to previous examples, except that we are not passing an image path on initialisation of :class:`aigarmic.Plate`. Instead, we are reading the image (using ``opencv``) and splitting (using :func:`aigarmic.process_plate_image.split_by_grid`) into small images. The key point here is that we then overwrite the attribute ``image_matrix`` of the :class:`aigarmic.Plate` object with the split images. Note also that we have to explicitly call :func:`aigarmic.Plate.annotate_images` after adding the plate image, since we did not provide the image on object construction. From this point onwards, the code is the same as before.
 
 To use a different splitting function, simply replace the call to :func:`aigarmic.process_plate_image.split_by_grid` with a custom function. The function must have a return type of ``list[list[ndarray]]``, i.e., a 2D matrix of images read using ``opencv``.
