### 2024-06-25 bb0bdcb
Create publish.yml
