+++ .github/workflows/publish.yml
@@ -1,0 +1,30 @@
+# Build the package and publish it to PyPI after tests pass.
+name: Publish to PyPI
+on:
+  push:
+    tags:
+      - "*"
+
+jobs:
+  tests:
+    uses: ./.github/workflows/pytest.yml
+  publish:
+    name: publish
+    needs: [tests] # require tests to pass before deploy runs
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout source
+        uses: actions/checkout@v3
+      - name: Set up Python 3.9
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.9
+      - name: Build package
+        run: |
+          python -m pip install -U pip build
+          python -m build
+      - name: Publish
+        uses: pypa/gh-action-pypi-publish@v1.5.0
+        with:
+          user: __token__
+          password: ${{ secrets.PYPI_KEY }}
