+++ .github/workflows/publish.yml
@@ -5,6 +5,11 @@ on:
     tags:
       - "*"
 
+permissions:
+  contents: write
+  checks: write
+  pull-requests: write
+
 jobs:
   tests:
     uses: ./.github/workflows/pytest.yml
