### 2024-06-25 agerada/AIgarMIC#11 @agerada
Yes that's right – the MIC can only be one of the `Plate` concentrations entered into the `PlateSet`. I think it will be interesting in future work to explore models that could perhaps calculate the MIC based on the image data provided (but without the MIC algorithm described here). In theory this could perhaps increase the resolution of the MIC (e.g., reporting between two measured concentrations) if a model is developed that demonstrates a relationship between the type of growth inhibition (e.g., faint film, single small colony, multiple small colonies), and MIC as measured by another method (broth microdilution).
