### 2024-06-25 agerada/AIgarMIC#16 @agerada
Thank you @gchure – happy to make all of those changes. I will work on them in the next few days.
