### 2024-06-26 agerada/AIgarMIC#11 @gchure
Okay, just wanted to make sure I was understanding it correctly. Great job on the documentation!
