### 2024-06-27 87d8715
remove default args for ncol/nrow in functions
