+++ tests/test_process_plate_image.py
@@ -11,20 +11,23 @@ def test_find_threshold_value(plates_images_paths, growth_image):
         image = cv2.imread(i)
         blur = cv2.GaussianBlur(image, (25, 25), 0)
         gray = cv2.cvtColor(blur, cv2.COLOR_BGR2GRAY)
-        thresholds, n = find_threshold_value(gray)
+        thresholds, n = find_threshold_value(image=gray, look_for=96)
         assert len(thresholds) == 96
 
     # fail, no grid on this image
     blur = cv2.GaussianBlur(growth_image, (25, 25), 0)
     gray = cv2.cvtColor(blur, cv2.COLOR_BGR2GRAY)
-    assert find_threshold_value(gray) is None
+    assert find_threshold_value(image=gray, look_for=96) is None
 
 
 @pytest.mark.assets_required
 def test_split_by_grid(plates_images_paths):
     for i in plates_images_paths:
         image = cv2.imread(i)
-        split_images = split_by_grid(image, visualise_contours=False)
+        split_images = split_by_grid(image,
+                                     n_rows=8,
+                                     n_cols=12,
+                                     visualise_contours=False)
         assert len(split_images) == 8
 
         # check that each row has 12 columns (inferred from rows)
