### 2024-07-01 e1912fe
command line example in paper
