### 2024-07-01 0f110f7
command line example in paper - fit to line
