### 2024-07-01 agerada/AIgarMIC#16 @agerada
Hi @gchure. Just a quick update on your last comment. I have made the proposed changes (see PR #30), look forward to see what you think.

As a side note, thank you for linking your recent paper and software (https://github.com/cremerlab/hplc-py). I have taken some more inspiration from your work, and integrated a workflow for publishing directly to pip (https://github.com/agerada/AIgarMIC/blob/main/.github/workflows/publish.yml). I'm very happy with the seamless testing and release workflow we have set up, which has made it easier to get releases out to users quicker.

Thanks again!
