### 2024-07-02 #6826 @ppxasjsm
Thank you @gchure and @agerada for all your work on the submission. I am sorry I missed some of the communication due to holidays and somehow not getting notifications. I think the paper submission here and the previous publication are ok. I'll now proceed with the editorial process.
