### 2024-07-02 #6826 @agerada
Hi @ppxasjsm - thank you for the update. The DOI suggested for `keras` does not appear to be a valid one for `keras`. I have double checked and both `keras` and `opencv` do not provide DOIs for citations.

I have also re-reviewed the proof and made some minor corrections https://github.com/agerada/AIgarMIC/pull/31 (mainly typos).
