### 2024-08-09 #6826 @agerada
Hi @ppxasjsm - just checking whether there there are any outsanding issues for me on this? If JOSS prefers all references to have an associated DOI, then I am happy to remove the references for `keras` and `opencv` and use hyperlinks to the packages within the text instead.
