### 2024-09-01 #6826 @ppxasjsm
@agerada can you also please provide a DOI for a Zenodo archive of the version of the software that is associated with this paper? Can you also post the version number here? The last thing I would like you to double-check is that all names and ORCIDs are correct.

I am really sorry about being quite slow with processing this over the summer but I ended up much busier than anticipated.
