### 2024-09-01 f70d74c
Small edits on grammar
