### 2024-09-02 89ec8b9
keras and opencv references changed to hyperlinks
