+++ paper.md
@@ -109,10 +109,10 @@ plate_set.generate_qc().tolist()
 
 ![`AIgarMIC` Plate and PlateSet API.\label{fig:api_plate}](paper_images/api_plate.png)
 
-In this example, images were not used -- growth codes were provided directly in matrix format. By providing images (imported using the `opencv` library [@itseez2015opencv]) to [`Plate`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/plate/index.html#aigarmic.plate.Plate) instances, `AIgarMIC` can automatically classify growth codes using a pre-trained model. `AIgarMIC` comes with a collection of [assets](https://datacat.liverpool.ac.uk/2631/) (example images and pre-trained models) to help
-users get started with the software [@geradaImageModelAssets2024]. Details of the built-in models, which are implemented as `keras` models (convolutional neural networks), can be found in the accompanying laboratory validation manuscript [@geradaDeterminationMinimumInhibitory2024; @chollet2015keras].
+In this example, images were not used -- growth codes were provided directly in matrix format. By providing images (imported using the [`opencv`](https://github.com/itseez/opencv) library to [`Plate`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/plate/index.html#aigarmic.plate.Plate) instances, `AIgarMIC` can automatically classify growth codes using a pre-trained model. `AIgarMIC` comes with a collection of [assets](https://datacat.liverpool.ac.uk/2631/) (example images and pre-trained models) to help
+users get started with the software [@geradaImageModelAssets2024]. Details of the built-in models, which are implemented as [`keras`](https://keras.io/) models (convolutional neural networks), can be found in the accompanying laboratory validation manuscript [@geradaDeterminationMinimumInhibitory2024].
 
-Alternatively, users can provide a custom model by inheriting from the base [`Model`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.Model) class (or the [`KerasModel`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.KerasModel) class if using a `keras` model) [@chollet2015keras]. Custom models must implement the `predict` method, which takes a colony image as input, and returns a `dictionary` containing, at a minimum, a `growth_code` member. \autoref{fig:api_model} shows the API for the [`Model`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.Model) class and subclasses. 
+Alternatively, users can provide a custom model by inheriting from the base [`Model`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.Model) class (or the [`KerasModel`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.KerasModel) class if using a [`keras`](https://keras.io/) model). Custom models must implement the `predict` method, which takes a colony image as input, and returns a `dictionary` containing, at a minimum, a `growth_code` member. \autoref{fig:api_model} shows the API for the [`Model`](https://aigarmic.readthedocs.io/en/latest/autoapi/aigarmic/model/index.html#aigarmic.model.Model) class and subclasses. 
 
[…]
