### 2024-09-02 5d9f004
replaced "test" with "assay" where appropriate, based on JOSS reviewer comments: https://github.com/agerada/AIgarMIC/issues/16#issue-2344986845
