### 2024-09-02 #6826 @agerada
Hi @ppxasjsm -- no need to apologise; I'm very grateful your editorial of our work, and for @gchure's constructive review.

I have merged https://github.com/agerada/AIgarMIC/pull/33 with a couple of minor edits from another read of the manuscript draft (@gchure had previously correctly flagged up that "assay" is a more apprpriate term than "test" with regards to the laboratory procdure of agar dilution MIC).

I can confirm that the author names and ORCIDs are correct.

Finally, here is a zenodo archive of the latest release (v1.1.1): https://zenodo.org/doi/10.5281/zenodo.13628882
