### 2024-09-02 #6826 @ppxasjsm
@editoribot set v1.1.1 as version
