### 2024-09-02 #6826 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
