### 2024-09-03 #6826 @agerada
Minor change to avoid breaking python statement over two pages: https://github.com/agerada/AIgarMIC/pull/35
