### 2024-09-05 #6826 @agerada
Thank you @Kevin-Mattheus-Moerman, I have made the changes (title and authors) to the ZENODO archive (https://zenodo.org/doi/10.5281/zenodo.13628882).
