### 2024-09-19 #6826 @agerada
Thank you @Kevin-Mattheus-Moerman. Just wanted to say a quick thank you to @ppxasjsm and @gchure for their input on this work and the very constuctive reviews. I look forward to contributing to JOSS in the future.
